openmax: Allow vendor OMX libraries to be loaded
Samsung OMX blobs extracted from a 5.0.1 stock ROM running on a
Galaxy Note 4 do not seem to have these components.
Not depending on them does not break anything either, so we remove
the checks here to allow the libs to be loaded by the open source
ExynosOMX_Core Component_Register class.

Change-Id: I8addef4d6772382d4ac3dafc818b63aaca8164e0